requirements.txt
asgiref==3.8.1
black==25.1.0
click==8.1.8
Django==5.1.4
django-htmx==1.21.0
gunicorn==23.0.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
sqlparse==0.5.3
whitenoise==6.9.0
